move beeswarms to just swarms
